refactor: work_page 내의 위젯 정리
- project.dart 파일로 프로젝트 분리
- project_detail_page로 프로젝트 내의 화면 분리
- work_page를 분리한 파일들을 조립하는 형태로 간소화
