+++ app/lib/src/pages/project.dart
@@ -1,0 +1,11 @@
+class Project {
+  final String title;
+  final DateTime createdAt;
+  final int practiceCount;
+
+  Project({
+    required this.title,
+    required this.createdAt,
+    required this.practiceCount,
+  });
+}
